add rangefeed level cleanup good addmuxer code
