+++ pkg/kv/kvserver/rangefeed/BUILD.bazel
@@ -9,10 +9,14 @@ go_library(
         "filter.go",
         "metrics.go",
         "processor.go",
+        "processor_test_helper.go",
         "registry.go",
+        "registry_test_helper.go",
         "resolved_timestamp.go",
         "scheduled_processor.go",
         "scheduler.go",
+        "stream_muxer.go",
+        "stream_muxer_helper.go",
         "task.go",
         "testutil.go",
     ],
@@ -47,6 +51,7 @@ go_library(
         "//pkg/util/timeutil",
         "//pkg/util/uuid",
         "@com_github_cockroachdb_errors//:errors",
+        "@com_github_stretchr_testify//require",
     ],
 )
 
@@ -63,6 +68,7 @@ go_test(
         "registry_test.go",
         "resolved_timestamp_test.go",
         "scheduler_test.go",
+        "stream_muxer_test.go",
         "task_test.go",
     ],
     embed = [":rangefeed"],
@@ -98,7 +104,7 @@ go_test(
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_cockroachdb_errors//oserror",
         "@com_github_cockroachdb_pebble//vfs",
-        "@com_github_stretchr_testify//assert",
         "@com_github_stretchr_testify//require",
+        "@com_github_zeebo_assert//:assert",
     ],
 )
