+++ pkg/kv/kvserver/rangefeed/registry.go
@@ -27,15 +27,15 @@ import (
 	"github.com/cockroachdb/errors"
 )
 
-// Stream is a object capable of transmitting RangeFeedEvents.
+// stream is a object capable of transmitting RangeFeedEvents.
 type Stream interface {
 	// Context returns the context for this stream.
 	Context() context.Context
 	// Send blocks until it sends m, the stream is done, or the stream breaks.
 	// Send must be safe to call on the same stream in different goroutines.
 	Send(*kvpb.RangeFeedEvent) error
-
 	Disconnect(err *kvpb.Error)
+	RegisterCleanUp(func())
 }
 
 // Shared event is an entry stored in registration channel. Each entry is
