+++ pkg/kv/kvserver/rangefeed/registry_test.go
@@ -13,9 +13,7 @@ package rangefeed
 import (
 	"context"
 	"fmt"
-	"sync"
 	"testing"
-	"time"
 
 	_ "github.com/cockroachdb/cockroach/pkg/keys" // hook up pretty printer
 	"github.com/cockroachdb/cockroach/pkg/kv/kvpb"
@@ -25,154 +23,9 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/testutils"
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
 	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
-	"github.com/cockroachdb/cockroach/pkg/util/syncutil"
 	"github.com/cockroachdb/cockroach/pkg/util/uuid"
 	"github.com/stretchr/testify/require"
 )
-
-var (
-	keyA, keyB = roachpb.Key("a"), roachpb.Key("b")
-	keyC, keyD = roachpb.Key("c"), roachpb.Key("d")
-	keyX, keyY = roachpb.Key("x"), roachpb.Key("y")
-
-	spAB = roachpb.Span{Key: keyA, EndKey: keyB}
-	spBC = roachpb.Span{Key: keyB, EndKey: keyC}
-	spCD = roachpb.Span{Key: keyC, EndKey: keyD}
-	spAC = roachpb.Span{Key: keyA, EndKey: keyC}
-	spXY = roachpb.Span{Key: keyX, EndKey: keyY}
-)
-
-type testStream struct {
-	ctx     context.Context
-	ctxDone func()
-	done    chan *kvpb.Error
-	mu      struct {
-		syncutil.Mutex
-		sendErr error
-		events  []*kvpb.RangeFeedEvent
-	}
-}
-
-func newTestStream() *testStream {
-	ctx, done := context.WithCancel(context.Background())
-	return &testStream{ctx: ctx, ctxDone: done, done: make(chan *kvpb.Error, 1)}
-}
-
-func (s *testStream) Context() context.Context {
-	return s.ctx
-}
-
-func (s *testStream) Cancel() {
-	s.ctxDone()
-}
-
-func (s *testStream) Send(e *kvpb.RangeFeedEvent) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if s.mu.sendErr != nil {
-		return s.mu.sendErr
-	}
-	s.mu.events = append(s.mu.events, e)
-	return nil
-}
-
-func (s *testStream) Disconnect(err *kvpb.Error) {
-	s.done <- err
-}
-
-func (s *testStream) Err(t *testing.T) error {
-	select {
-	case err := <-s.done:
-		return err.GoError()
-	case <-time.After(30 * time.Second):
-		t.Fatalf("time out waiting for rangefeed completion")
-		return nil
-	}
-}
-
-func (s *testStream) SetSendErr(err error) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.mu.sendErr = err
-}
-
-func (s *testStream) Events() []*kvpb.RangeFeedEvent {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	es := s.mu.events
-	s.mu.events = nil
-	return es
-}
-
-func (s *testStream) BlockSend() func() {
-	s.mu.Lock()
-	var once sync.Once
-	return func() {
-		once.Do(s.mu.Unlock) // safe to call multiple times, e.g. defer and explicit
-	}
-}
-
-type testRegistration struct {
-	registration
-	stream *testStream
-}
-
-func makeCatchUpIterator(
-	iter storage.SimpleMVCCIterator, span roachpb.Span, startTime hlc.Timestamp,
-) *CatchUpIterator {
-	if iter == nil {
-		return nil
-	}
-	return &CatchUpIterator{
-		simpleCatchupIter: simpleCatchupIterAdapter{iter},
-		span:              span,
-		startTime:         startTime,
-	}
-}
-
-func newTestRegistration(
-	span roachpb.Span,
-	ts hlc.Timestamp,
-	catchup storage.SimpleMVCCIterator,
-	withDiff bool,
-	withFiltering bool,
-) *testRegistration {
-	s := newTestStream()
-	r := newRegistration(
-		span,
-		ts,
-		makeCatchUpIterator(catchup, span, ts),
-		withDiff,
-		withFiltering,
-		5,
-		false, /* blockWhenFull */
-		NewMetrics(),
-		s,
-		func() {},
-	)
-	return &testRegistration{
-		registration: r,
-		stream:       s,
-	}
-}
-
-func (r *testRegistration) Events() []*kvpb.RangeFeedEvent {
-	return r.stream.Events()
-}
-
-func (r *testRegistration) Err() error {
-	err := <-r.stream.done
-	return err.GoError()
-}
-
-func (r *testRegistration) TryErr() error {
-	select {
-	case err := <-r.stream.done:
-		return err.GoError()
-	default:
-		return nil
-	}
-}
 
 func TestRegistrationBasic(t *testing.T) {
 	defer leaktest.AfterTest(t)()
@@ -191,7 +44,7 @@ func TestRegistrationBasic(t *testing.T) {
 	require.Equal(t, len(noCatchupReg.buf), 2)
 	go noCatchupReg.runOutputLoop(ctx, 0)
 	require.NoError(t, noCatchupReg.waitForCaughtUp(ctx))
-	require.Equal(t, []*kvpb.RangeFeedEvent{ev1, ev2}, noCatchupReg.stream.Events())
+	require.Equal(t, []*kvpb.RangeFeedEvent{ev1, ev2}, noCatchupReg.testStream.Events())
 	noCatchupReg.disconnect(nil)
 
 	// Registration with catchup scan.
@@ -207,7 +60,7 @@ func TestRegistrationBasic(t *testing.T) {
 	require.Equal(t, len(catchupReg.buf), 2)
 	go catchupReg.runOutputLoop(ctx, 0)
 	require.NoError(t, catchupReg.waitForCaughtUp(ctx))
-	events := catchupReg.stream.Events()
+	events := catchupReg.testStream.Events()
 	require.Equal(t, 5, len(events))
 	require.Equal(t, []*kvpb.RangeFeedEvent{ev1, ev2}, events[3:])
 	catchupReg.disconnect(nil)
@@ -222,8 +75,8 @@ func TestRegistrationBasic(t *testing.T) {
 	require.NoError(t, disconnectReg.waitForCaughtUp(ctx))
 	discErr := kvpb.NewError(fmt.Errorf("disconnection error"))
 	disconnectReg.disconnect(discErr)
-	require.Equal(t, discErr.GoError(), disconnectReg.Err())
-	require.Equal(t, 2, len(disconnectReg.stream.Events()))
+	require.Equal(t, discErr.GoError(), disconnectReg.Err(t))
+	require.Equal(t, 2, len(disconnectReg.testStream.Events()))
 
 	// External Disconnect before output loop.
 	disconnectEarlyReg := newTestRegistration(spAB, hlc.Timestamp{}, nil, /* catchup */
@@ -232,8 +85,8 @@ func TestRegistrationBasic(t *testing.T) {
 	disconnectEarlyReg.publish(ctx, ev2, nil /* alloc */)
 	disconnectEarlyReg.disconnect(discErr)
 	go disconnectEarlyReg.runOutputLoop(ctx, 0)
-	require.Equal(t, discErr.GoError(), disconnectEarlyReg.Err())
-	require.Equal(t, 0, len(disconnectEarlyReg.stream.Events()))
+	require.Equal(t, discErr.GoError(), disconnectEarlyReg.Err(t))
+	require.Equal(t, 0, len(disconnectEarlyReg.testStream.Events()))
 
 	// Overflow.
 	overflowReg := newTestRegistration(spAB, hlc.Timestamp{}, nil, /* catchup */
@@ -242,25 +95,26 @@ func TestRegistrationBasic(t *testing.T) {
 		overflowReg.publish(ctx, ev1, nil /* alloc */)
 	}
 	go overflowReg.runOutputLoop(ctx, 0)
-	require.Equal(t, newErrBufferCapacityExceeded().GoError(), overflowReg.Err())
+	require.Equal(t, newErrBufferCapacityExceeded().GoError(), overflowReg.Err(t))
 	require.Equal(t, cap(overflowReg.buf), len(overflowReg.Events()))
 
 	// Stream Error.
 	streamErrReg := newTestRegistration(spAB, hlc.Timestamp{}, nil, /* catchup */
 		false /* withDiff */, false /* withFiltering */)
 	streamErr := fmt.Errorf("stream error")
-	streamErrReg.stream.SetSendErr(streamErr)
+
+	streamErrReg.testStream.SetSendErr(streamErr)
 	go streamErrReg.runOutputLoop(ctx, 0)
 	streamErrReg.publish(ctx, ev1, nil /* alloc */)
-	require.Equal(t, streamErr.Error(), streamErrReg.Err().Error())
+	require.Equal(t, streamErr.Error(), streamErrReg.Err(t).Error())
 
 	// Stream Context Canceled.
 	streamCancelReg := newTestRegistration(spAB, hlc.Timestamp{}, nil, /* catchup */
 		false /* withDiff */, false /* withFiltering */)
-	streamCancelReg.stream.Cancel()
+	streamCancelReg.stream.Disconnect(kvpb.NewError(context.Canceled))
 	go streamCancelReg.runOutputLoop(ctx, 0)
 	require.NoError(t, streamCancelReg.waitForCaughtUp(ctx))
-	require.Equal(t, streamCancelReg.stream.Context().Err(), streamCancelReg.Err())
+	require.Equal(t, streamCancelReg.stream.Context().Err(), streamCancelReg.Err(t))
 }
 
 func TestRegistrationCatchUpScan(t *testing.T) {
@@ -484,7 +338,7 @@ func TestRegistryBasic(t *testing.T) {
 	// Disconnect span that overlaps with rCD.
 	reg.DisconnectWithErr(ctx, spCD, err1)
 	require.Equal(t, 4, reg.Len())
-	require.Equal(t, err1.GoError(), rCD.Err())
+	require.Equal(t, err1.GoError(), rCD.Err(t))
 
 	// Can still publish to rAB.
 	reg.PublishToOverlapping(ctx, spAB, ev4, false /* omitInRangefeeds */, nil /* alloc */)
@@ -496,8 +350,8 @@ func TestRegistryBasic(t *testing.T) {
 
 	// Disconnect from rAB without error.
 	reg.Disconnect(ctx, spAB)
-	require.Nil(t, rAC.Err())
-	require.Nil(t, rAB.Err())
+	require.Nil(t, rAC.Err(t))
+	require.Nil(t, rAB.Err(t))
 	require.Equal(t, 1, reg.Len())
 
 	// Check the registry's operation filter again.
