+++ pkg/server/BUILD.bazel
@@ -38,7 +38,6 @@ go_library(
         "migration.go",
         "node.go",
         "node_http_router.go",
-        "node_stream_muxer.go",
         "node_tenant.go",
         "node_tombstone_storage.go",
         "nodes_response.go",
@@ -436,7 +435,6 @@ go_test(
         "migration_test.go",
         "multi_store_test.go",
         "node_http_router_test.go",
-        "node_stream_muxer_test.go",
         "node_tenant_test.go",
         "node_test.go",
         "node_tombstone_storage_test.go",
@@ -555,7 +553,6 @@ go_test(
         "//pkg/util/protoutil",
         "//pkg/util/randutil",
         "//pkg/util/stop",
-        "//pkg/util/syncutil",
         "//pkg/util/timeutil",
         "//pkg/util/tracing",
         "//pkg/util/tracing/tracingpb",
